Move knowledge graph to nidx_tests for reusability
